% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCD.R
\name{readCD}
\alias{readCD}
\title{Read precipitation and temperature data for quality controls}
\usage{
readCD(x)
}
\arguments{
\item{x}{A string name file}
}
\value{
A tibble with columns yy (year), mm (month), dd (day) and as many columns as the number of stations
}
\description{
Read precipitation and temperature data for quality controls
}
